+++ examples/basic_ibe/frontend/scripts/gen_bindings.sh
@@ -1,11 +1,10 @@
 #!/bin/bash
 
-cd ../../backend && make extract-candid
+cd ../../rust/backend && make extract-candid
 
-cd .. && dfx generate basic_ibe || exit 1
+cd ../..
 
-rm -r frontend/src/declarations/basic_ibe > /dev/null 2>&1 || true
+rm -rf frontend/src/declarations/basic_ibe
 
 mkdir -p frontend/src/declarations/basic_ibe
-mv src/declarations/basic_ibe frontend/src/declarations
-rmdir -p src/declarations > /dev/null 2>&1 || true
+npx @icp-sdk/bindgen --did-file rust/backend/backend.did --out-dir frontend/src/declarations/basic_ibe --declarations-flat --force