+++ examples/encrypted_notes_dapp_vetkd/frontend/src/components/Notes.svelte
@@ -2,6 +2,7 @@
   import type { NoteModel } from '../lib/note';
 
   import { notesStore } from '../store/notes';
+  import { link } from 'svelte-spa-router';
   import Header from './Header.svelte';
   import Note from './Note.svelte';
   import Spinner from './Spinner.svelte';
@@ -36,7 +37,7 @@
   <span slot="title"> Your notes </span>
   <svelte:fragment slot="actions">
     {#if $notesStore.state === 'loaded' && $notesStore.list.length > 0}
-      <a class="btn btn-primary" href="/">New Note</a>
+      <a class="btn btn-primary" href="/" use:link>New Note</a>
     {/if}
   </svelte:fragment>
 </Header>
@@ -66,7 +67,7 @@
     {:else}
       <div class="text-center pt-8 italic">You don't have any notes.</div>
       <div class="text-center pt-8 ">
-        <a href="/" class="btn btn-primary">Add a note</a>
+        <a href="/" class="btn btn-primary" use:link>Add a note</a>
       </div>
     {/if}
   {:else if $notesStore.state === 'error'}